refactor: add a file to unit fakeData
